[hironx py] Let setTargetPose fail when invalid kinematic group name passed.
**Background**
(Same commit message as https://github.com/fkanehiro/hrpsys-base/pull/1113)
`setTargetPose` takes a kinematic group as its 1st arg. When invalid value passed, it still tries to solve IK and always fails with an IK solution failure message, which is ambiguous because it failed not because
the computation but because the target group name is simply invalid.

**Approach to fix**
With this change it first checks if the passed value is a registered kinematic group name and if not it fails with the reason accordingly.

**Additional note at rtmros_hironx**
Same change as https://github.com/fkanehiro/hrpsys-base/pull/1113. This method should be removed as part of https://github.com/start-jsk/rtmros_hironx/pull/470, once https://github.com/fkanehiro/hrpsys-base/pull/1063 resolves.